Add shadow effect to LightingEffect class
