+++ docs/api-reference/lights/directional-light.md
@@ -33,6 +33,7 @@ const directionalLight = new DirectionalLight({color, intensity, direction});
 * `color` - (*array*,)  RGB color of directional light source, default value is `[255, 255, 255]`.
 * `intensity` - (*number*) Strength of directional light source, default value is `1.0`.
 * `direction` - (*array*,)  3D vector specifies the direction the light comes from, default value is `[0, 0, -1]`.
+* `castShadow` - (*boolean*, optional) Enable experimental shadow effect, default value is `false`.
 
 ## Source
